Add files via upload
For Task 3 of my Oasis Infobyte internship, I developed a Temperature Converter web application. The goal of this project was to create a simple and user-friendly interface that allows users to input temperatures in Celsius, Fahrenheit, or Kelvin, and instantly convert them to the other units. The application includes an input validation feature to ensure correct numeric input and provides clear conversion results upon clicking the "Convert" button. This project helped me enhance my skills in JavaScript, HTML, and CSS, and provided hands-on experience in building interactive web applications.